Apply ContainerLifecycleHooks to MongoDBContainer
